Move tests from CoreCliTest to PmdCliTest related to --relativize-paths-with
