+++ pmd-cli/src/test/java/net/sourceforge/pmd/cli/PmdCliTest.java
@@ -7,21 +7,24 @@ package net.sourceforge.pmd.cli;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.containsString;
 import static org.hamcrest.Matchers.containsStringIgnoringCase;
+import static org.hamcrest.Matchers.endsWith;
 import static org.hamcrest.Matchers.not;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
+import java.io.File;
 import java.io.IOException;
 import java.nio.ByteBuffer;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.FileSystems;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.List;
 
-import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
@@ -46,8 +49,8 @@ class PmdCliTest extends BaseCliTest {
 
     private Path srcDir;
 
-    @AfterAll
-    static void resetLogging() {
+    @AfterEach
+    void resetLogging() {
         // reset logging in case "--debug" changed the logging properties
         // See also Slf4jSimpleConfigurationForAnt
         Slf4jSimpleConfiguration.reconfigureDefaultLogLevel(null);
@@ -212,11 +215,121 @@ class PmdCliTest extends BaseCliTest {
     }
 
     @Test
+    void testNoRelativizeWithAbsoluteSrcDir() throws Exception {
+        assertTrue(srcDir.isAbsolute(), "srcDir should be absolute");
+        String log = runCli(ExecutionResult.VIOLATIONS_FOUND, "--dir", srcDir.toString(), "--rulesets",
+                DUMMY_RULESET_WITH_VIOLATIONS);
+
+        assertThat(log, containsString(srcDir.resolve("someSource.dummy").toString()));
+    }
+
+    @Test
+    void testNoRelativizeWithRelativeSrcDir() throws Exception {
+        // Note, that we can't reliably change the current working directory for the current java process
+        // therefore we use the current directory and make sure, we are at the correct place - in pmd-core
+        Path cwd = Paths.get(".").toRealPath();
+        assertThat(cwd.toString(), endsWith("pmd-cli"));
+        String relativeSrcDir = "src/test/resources/net/sourceforge/pmd/cli/src";
+        assertTrue(Files.isDirectory(cwd.resolve(relativeSrcDir)));
+
+        String log = runCli(ExecutionResult.VIOLATIONS_FOUND, "--dir", relativeSrcDir, "--rulesets",
+                DUMMY_RULESET_WITH_VIOLATIONS);
+
+        assertThat(log, containsString("\n" + IOUtil.normalizePath(relativeSrcDir + "/somefile.dummy")));
+    }
+
+    @Test
+    void testNoRelativizeWithRelativeSrcDirParent() throws Exception {
+        // Note, that we can't reliably change the current working directory for the current java process
+        // therefore we use the current directory and make sure, we are at the correct place - in pmd-core
+        Path cwd = Paths.get(".").toRealPath();
+        assertThat(cwd.toString(), endsWith("pmd-cli"));
+        String relativeSrcDir = IOUtil.normalizePath("src/test/resources/net/sourceforge/pmd/cli/src");
+        assertTrue(Files.isDirectory(cwd.resolve(relativeSrcDir)));
+
+        // use the parent directory
+        String relativeSrcDirWithParent = relativeSrcDir + File.separator + "..";
+
+        String log = runCli(ExecutionResult.VIOLATIONS_FOUND, "--dir", relativeSrcDirWithParent, "--rulesets",
+                DUMMY_RULESET_WITH_VIOLATIONS);
+
+        assertThat(log, containsString("\n" + relativeSrcDirWithParent + IOUtil.normalizePath("/src/somefile.dummy")));
+    }
+
+    @Test
+    void testRelativizeWithRootRelativeSrcDir() throws Exception {
+        // Note, that we can't reliably change the current working directory for the current java process
+        // therefore we use the current directory and make sure, we are at the correct place - in pmd-core
+        Path cwd = Paths.get(".").toRealPath();
+        assertThat(cwd.toString(), endsWith("pmd-cli"));
+        String relativeSrcDir = "src/test/resources/net/sourceforge/pmd/cli/src";
+        assertTrue(Files.isDirectory(cwd.resolve(relativeSrcDir)));
+
+        String root = cwd.getRoot().toString();
+
+        String log = runCli(ExecutionResult.VIOLATIONS_FOUND, "--dir", relativeSrcDir, "--rulesets",
+                DUMMY_RULESET_WITH_VIOLATIONS, "--relativize-paths-with", root);
+
+        String absoluteSrcPath = cwd.resolve(relativeSrcDir).resolve("somefile.dummy").toString();
+        assertThat(log, containsString("\n" + absoluteSrcPath));
+    }
+
+    @Test
     void testRelativizeWith() throws Exception {
         final String log = runCli(ExecutionResult.VIOLATIONS_FOUND, "--dir", srcDir.toString(), "--rulesets",
                 DUMMY_RULESET_WITH_VIOLATIONS, "-z", srcDir.getParent().toString());
         assertThat(log, not(containsString(srcDir.resolve("someSource.dummy").toString())));
         assertThat(log, containsString("\n" + IOUtil.normalizePath("src/someSource.dummy")));
+    }
+
+    @Test
+    void testRelativizeWithSymLink() throws Exception {
+        // srcDir = /tmp/junit123/src
+        // symlinkedSrcDir = /tmp/junit123/sources -> /tmp/junit123/src
+        Path symlinkedSrcDir = Files.createSymbolicLink(tempRoot().resolve("sources"), srcDir);
+        String log = runCli(ExecutionResult.VIOLATIONS_FOUND, "--dir", symlinkedSrcDir.toString(), "--rulesets",
+                DUMMY_RULESET_WITH_VIOLATIONS, "-z", symlinkedSrcDir.toString());
+
+        assertThat(log, not(containsString(srcDir.resolve("someSource.dummy").toString())));
+        assertThat(log, not(containsString(symlinkedSrcDir.resolve("someSource.dummy").toString())));
+        assertThat(log, containsString("\nsomeSource.dummy"));
+    }
+
+    @Test
+    void testRelativizeWithSymLinkParent() throws Exception {
+        // srcDir = /tmp/junit123/src
+        // symlinkedSrcDir = /tmp/junit-relativize-with-123 -> /tmp/junit123/src
+        Path tempPath = Files.createTempDirectory("junit-relativize-with-");
+        Files.delete(tempPath);
+        Path symlinkedSrcDir = Files.createSymbolicLink(tempPath, srcDir);
+        // relativizing against parent of symlinkedSrcDir: /tmp
+        String log = runCli(ExecutionResult.VIOLATIONS_FOUND, "--dir", symlinkedSrcDir.toString(), "--rulesets",
+                DUMMY_RULESET_WITH_VIOLATIONS, "-z", symlinkedSrcDir.getParent().toString());
+
+        assertThat(log, not(containsString(srcDir.resolve("someSource.dummy").toString())));
+        assertThat(log, not(containsString(symlinkedSrcDir.resolve("someSource.dummy").toString())));
+        // base path is symlinkedSrcDir without /tmp: e.g. junit-relativize-with-123
+        String basePath = symlinkedSrcDir.getParent().relativize(symlinkedSrcDir).toString();
+        assertThat(log, containsString("\n" + basePath + File.separator + "someSource.dummy"));
+    }
+
+    @Test
+    void testRelativizeWithMultiple() throws Exception {
+        String log = runCli(ExecutionResult.VIOLATIONS_FOUND, "--dir", srcDir.toString(), "--rulesets",
+                DUMMY_RULESET_WITH_VIOLATIONS, "-z", srcDir.getParent().toString(), "-z", srcDir.toString());
+
+        assertThat(log, not(containsString(srcDir.resolve("someSource.dummy").toString())));
+        assertThat(log, containsString("\nsomeSource.dummy"));
+    }
+
+    @Test
+    void testRelativizeWithFileIsError() throws Exception {
+        String log = runCli(ExecutionResult.USAGE_ERROR, "--dir", srcDir.toString(), "--rulesets",
+                DUMMY_RULESET_WITH_VIOLATIONS, "-z", srcDir.resolve("someSource.dummy").toString());
+
+        assertThat(log, containsString(
+                "Expected a directory path for option '--relativize-paths-with', found a file: "
+                        + srcDir.resolve("someSource.dummy")));
     }
 
     // utilities
