+++ pmd-cli/src/test/resources/net/sourceforge/pmd/cli/src/anotherfile.dummy
@@ -1,0 +1,1 @@
+Another file for testing
