+++ pmd-cli/src/test/resources/net/sourceforge/pmd/cli/src/somefile3.dummy
@@ -1,0 +1,1 @@
+Some file for testing
