--- pmd-core/src/test/resources/net/sourceforge/pmd/cli/FakeRuleset2.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0"?>
-<ruleset name="Test Ruleset" xmlns="http://pmd.sourceforge.net/ruleset/2.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://pmd.sourceforge.net/ruleset/2.0.0 https://pmd.sourceforge.io/ruleset_2_0_0.xsd">
-
-    <description>
-  Ruleset used by test RuleSetFactoryTest
-  </description>
-
-    <rule name="Ruleset3Rule1" language="dummy" since="1.0" message="Test Rule 1" class="net.sourceforge.pmd.cli.CoreCliTest$FooRule"
-        externalInfoUrl="${pmd.website.baseurl}/rules/test/TestRuleset3.xml#Ruleset3Rule1">
-        <description>
-Just for test
-     </description>
-        <priority>3</priority>
-        <example>
- <![CDATA[
- ]]>
-     </example>
-    </rule>
-
-</ruleset>
